+++ scripts/transactions/pyth/schema.ts
